@@ -1,59 +1,49 @@
 {
-  "name": "client",
-  "version": "0.1.0",
+  "name": "universal-directory-client",
   "private": true,
-  "dependencies": {
-    "@tailwindcss/forms": "^0.5.7",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.4.0",
-    "blueimp-md5": "^2.19.0",
-    "classnames": "^2.3.2",
-    "firebase": "^10.1.0",
-    "query-string": "^8.1.0",
-    "ramda": "^0.30.1",
-    "react": "^18.2.0",
-    "react-cookie": "^7.2.2",
-    "react-dom": "^18.2.0",
-    "react-flatpickr": "^3.10.13",
-    "react-router-dom": "^5.3.4",
-    "react-scripts": "5.0.1",
-    "react-spinners": "^0.13.8",
-    "react-toastify": "^9.1.3",
-    "react-transition-group": "^4.4.5",
-    "url": "^0.11.3",
-    "web-vitals": "^2.1.4"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "lint": "eslint ./src",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@tailwindcss/vite": "^4.1.4",
+    "axios": "^1.8.4",
+    "blueimp-md5": "^2.19.0",
+    "classnames": "^2.5.1",
+    "firebase": "^11.6.0",
+    "prop-types": "^15.8.1",
+    "ramda": "^0.30.1",
+    "react": "^19.0.0",
+    "react-cookie": "^8.0.1",
+    "react-dom": "^19.0.0",
+    "react-flatpickr": "^4.0.4",
+    "react-router-dom": "^7.5.1",
+    "react-spinners": "^0.16.1",
+    "react-toastify": "^11.0.5",
+    "react-transition-group": "^4.4.5"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "tailwindcss": "^3.4.1"
-  },
-  "proxy": "https://gw-oauth-server-hcf3ceajdpg2gcbg.canadacentral-01.azurewebsites.net/",
-  "home": "."
+    "@eslint/js": "^9.22.0",
+    "@tailwindcss/postcss": "^4.1.4",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "@vitejs/plugin-react": "^4.3.4",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.22.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^16.0.0",
+    "jsdom": "^26.1.0",
+    "postcss": "^8.5.3",
+    "tailwindcss": "^4.1.4",
+    "vite": "^6.3.1",
+    "vitest": "^3.1.1"
+  }
 }
