@@ -1,8 +1,6 @@
-module.exports = ({ env }) => ({
-  plugins: [
-    require('tailwindcss')({
-      config: './tailwind.config.js'
-    }),
-    require('autoprefixer')()
-  ],
-})
+import tailwindcss from "@tailwindcss/postcss";
+import autoprefixer from "autoprefixer";
+
+export default {
+  plugins: [tailwindcss(), autoprefixer()],
+};
