@@ -1,6 +1,6 @@
 import React from "react";
 import { useEffect, useState } from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 import { useAuth } from "../../component/AuthContext";
 import Toolbox from "../../component/Toolbox";
@@ -11,10 +11,10 @@ import apiApi from "../../api/api-api";
 import domainApi from "../../api/domain-api";
 
 function ApiPage({ status }) {
-  const history = useHistory();
+  const navigate = useNavigate();
   const pageDisplayCount = 10;
   const postDisplayCount = 15;
-  const { accessToken, company, domain } = useAuth();
+  const { company, domain } = useAuth();
   const [apis, setApis] = useState([]);
   const [domainId, setDomainId] = useState(domain.id);
   const [checkedItems, setCheckedItems] = useState([]);
@@ -42,15 +42,17 @@ function ApiPage({ status }) {
     setCheckedItems(childCheckedItems);
   };
 
-  const onClickNew = function (e) {
-    history.push("/apis-new", {
-      company: company,
-      domain: domain,
-      mode: "new",
+  const onClickNew = function () {
+    navigate("/apis-new", {
+      state: {
+        company: company,
+        domain: domain,
+        mode: "new",
+      },
     });
   };
 
-  const onClickDel = async function (e) {
+  const onClickDel = async function () {
     //const api = await apiApi.get(company.id, domain.id, checkedItems);
     await apiApi.remove(checkedItems);
     setCheckedItems([]);
@@ -58,20 +60,24 @@ function ApiPage({ status }) {
   };
 
   const onClickView = (item) => {
-    history.push("/apis-view", {
-      company: company,
-      domain: domain,
-      api: item,
-      mode: "view",
+    navigate("/apis-view", {
+      state: {
+        company: company,
+        domain: domain,
+        api: item,
+        mode: "view",
+      },
     });
   };
 
   const onClickEdit = (item) => {
-    history.push("/apis-new", {
-      company: company,
-      domain: domain,
-      api: item,
-      mode: "edit",
+    navigate("/apis-new", {
+      state: {
+        company: company,
+        domain: domain,
+        api: item,
+        mode: "edit",
+      },
     });
   };
 
@@ -80,9 +86,9 @@ function ApiPage({ status }) {
   };
 
   const getApis = async (domId) => {
-    if (accessToken) {
-      const condition = ["companyId", "==", company.id];
+    try {
       const items = await apiApi.get(company.id, domId, null, null);
+
       setApis(items.data);
       setPageEnd(
         Math.ceil(items.data.length / postsPerPage) < pageDisplayCount
@@ -95,6 +101,8 @@ function ApiPage({ status }) {
         setDomains(doms.data);
       }
       setDomainId(domId);
+    } catch (error) {
+      if (error.status === 401) navigate("/");
     }
   };
 
@@ -103,31 +111,31 @@ function ApiPage({ status }) {
     getApis(domainId);
     setDelBtnLabel(status == "active" ? "Delete" : "Recover");
     setIsLoading(false);
-  }, [domains]);
+  }, [domainId]);
 
-  const delSvg = (
-    <svg
-      class="w-4 h-4 mr-2"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      stroke-width="2"
-      stroke="currentColor"
-      fill="none"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    >
-      {" "}
-      <path stroke="none" d="M0 0h24v24H0z" />{" "}
-      <line x1="4" y1="7" x2="20" y2="7" />{" "}
-      <line x1="10" y1="11" x2="10" y2="17" />{" "}
-      <line x1="14" y1="11" x2="14" y2="17" />{" "}
-      <path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" />{" "}
-      <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" />
-    </svg>
-  );
-  const purgeClass =
-    "w-30 ml-8 bg-gray-300 disabled:hover:cursor-not-allowed disabled:bg-gray-100 disabled:text-gray-400 enabled:transition enabled:transform enabled:hover:translate-x-1 enabled:hover:bg-gray-400 text-gray-800 font-bold py-2 px-4 rounded inline-flex items-center";
+  // const delSvg = (
+  //   <svg
+  //     className="w-4 h-4 mr-2"
+  //     width="24"
+  //     height="24"
+  //     viewBox="0 0 24 24"
+  //     strokeWidth="2"
+  //     stroke="currentColor"
+  //     fill="none"
+  //     strokeLinecap="round"
+  //     strokeLinejoin="round"
+  //   >
+  //     {" "}
+  //     <path stroke="none" d="M0 0h24v24H0z" />{" "}
+  //     <line x1="4" y1="7" x2="20" y2="7" />{" "}
+  //     <line x1="10" y1="11" x2="10" y2="17" />{" "}
+  //     <line x1="14" y1="11" x2="14" y2="17" />{" "}
+  //     <path d="M5 7l1 12a2 2 0 0 0 2 2h8a2 2 0 0 0 2 -2l1 -12" />{" "}
+  //     <path d="M9 7v-3a1 1 0 0 1 1 -1h4a1 1 0 0 1 1 1v3" />
+  //   </svg>
+  // );
+  // const purgeClass =
+  //   "w-30 ml-8 bg-gray-300 disabled:hover:cursor-not-allowed disabled:bg-gray-100 disabled:text-gray-400 enabled:transition enabled:transform enabled:hover:translate-x-1 enabled:hover:bg-gray-400 text-gray-800 font-bold py-2 px-4 rounded inline-flex items-center";
 
   return (
     <div className="col-span-full xl:col-span-6 bg-white dark:bg-slate-800 shadow-lg rounded-sm border border-slate-200 dark:border-slate-700">
