@@ -1,38 +1,39 @@
 import { useEffect, useState } from "react";
-import { useLocation, useHistory } from "react-router-dom";
-import { encodeClientCredentials } from "../utils/Utils";
+import { useLocation, useNavigate } from "react-router-dom";
+import { encodeClientCredentials, generateCodeVerifier } from "../utils/Utils";
 import axios from "axios";
 import { useAuth } from "../component/AuthContext";
-import { client, authServer } from "../api/igw-api";
-import { uniDirServer } from "../api/igw-api";
+import { client, authServer, uniDirServer } from "../api/igw-api";
 import ButtonToolbox from "../component/ButtonToolbox";
 import AuthFormHeader from "../component/AuthFormHeader";
 import { httpClient, setHttpClient } from "../api/httpClient";
 import { useCookies } from "react-cookie";
 
-export default function AuthCallback(props) {
+export default function AuthCallback() {
   const location = useLocation();
-  const history = useHistory();
+  const navigate = useNavigate();
   const [title, setTitle] = useState("Logging.....");
   const {
-    user,
     saveUser,
     saveClient,
     saveCompany,
     saveDomain,
-    saveRootCompany,
-    saveAccessToken,
+    codeVerifier,
+    saveCodeVerifier,
   } = useAuth();
   const searchParams = new URLSearchParams(location.search);
+
   const code = searchParams.get("code");
   const error = searchParams.get("error");
-  const [cookies, setCookie] = useCookies(["igw-udir"]);
-
+  const [cookies, _] = useCookies(["igw-udir"]);
   const form_data = {
     grant_type: "authorization_code",
     code: code,
     redirect_uri: client.redirect_uris[0],
+    code_verifier: codeVerifier,
   };
+
+  saveCodeVerifier(null);
 
   const headers = {
     "Content-Type": "application/x-www-form-urlencoded",
@@ -44,70 +45,102 @@ export default function AuthCallback(props) {
   useEffect(() => {
     (async function () {
       try {
-        if (cookies.tokenJson) {
-          await fetchDashboard(cookies.tokenJson);
-          history.push("/dashboard");
-        } else if (code && !user) {
+        // if (cookies.tokenJson) {
+        //   await fetchDashboard(cookies.tokenJson);
+        //   navigate("/dashboard");
+        // } else
+        if (code) {
           const axiosAuth = axios.create();
           const tokenJson = await axiosAuth.post(
             authServer.tokenEndpoint,
             form_data,
-            { headers: headers }
+            { withCredentials: true, headers: headers }
           );
+          // const userSession = {id:tokenJson.data.user.id, email:tokenJson.data.user.email, root:tokenJson.data.user.root};
+          // saveUser(userSession);
 
-          saveUser(tokenJson.data.user);
-          //saveAccessToken(tokenJson.data.access_token);
-          saveClient(tokenJson.data.client);
-
-          if (cookies.IsRemember) {
-            setCookie("tokenJson", tokenJson, {
-              Expires: new Date(Date.now() + 3600 * 1000), // expires in 1 hour
-              Secure: true, // cookie will only be sent over HTTPS
-              SameSite: "Strict", // prevent cross-site request
-              //httpOnly: true
-            });
-          }
+          //saveClient(tokenJson.data.client);
           if (tokenJson.data.access_token) {
-            await fetchDashboard(tokenJson);
+            await fetchDashboard(tokenJson, cookies.IsRemember);
           }
 
-          history.push("/dashboard");
+          navigate("/dashboard");
         }
       } catch (error) {
-        history.push("/error");
+        navigate("/error");
         console.log("error", error);
       }
     })();
   }, []);
 
   const fetchDashboard = async (tokenJson) => {
-    saveUser(tokenJson.data.user);
-    saveAccessToken(tokenJson.data.access_token);
-    saveClient(tokenJson.data.client);
+    const clientSession = {
+      companyId: tokenJson.data.client.companyId,
+      domainId: tokenJson.data.client.domain,
+    };
+    saveClient(clientSession);
 
     if (tokenJson.data.access_token) {
       const header = {
         "Access-Control-Allow-Origin": "*",
         Authorization: `Bearer ${tokenJson.data.access_token}`,
       };
-      //saveHeader(header);
+
       setHttpClient(header);
-      //const com = await axiosAPI.get(`${uniDirServer.Endpoint}/companys/${tokenJson.data.client.companyId}`, {headers: header});
       const com = await httpClient.get(
-        `${uniDirServer.Endpoint}/companys/${tokenJson.data.client.companyId}`
+        `${uniDirServer.Endpoint}/companys/${tokenJson.data.user.companyId}`
       );
-      saveCompany(com.data);
-      saveRootCompany(com.data);
-      //const dom = await axiosAPI.get(`${uniDirServer.Endpoint}/companys/${tokenJson.data.client.companyId}/domainNames/${tokenJson.data.client.domain}`, {headers: header});
+
+      const companySession = {
+        id: com.data.id,
+        name: com.data.name,
+        displayName: com.data.displayName,
+        parent: com.data.parent,
+        type: com.data.type ? com.data.type : "customer",
+      };
+      saveCompany(companySession);
+      //saveRootCompany(com.data);
+
       const dom = await httpClient.get(
-        `${uniDirServer.Endpoint}/companys/${tokenJson.data.client.companyId}/domainNames/${tokenJson.data.client.domain}`
+        `${uniDirServer.Endpoint}/companys/${tokenJson.data.client.companyId}/domainNames/${tokenJson.data.user.domainId}`
       );
-      saveDomain(dom.data);
+      const domainSession = {
+        id: dom.data.id,
+        name: dom.data.name,
+        description: dom.data.description,
+      };
+      saveDomain(domainSession);
+
+      const userSession = {
+        id: tokenJson.data.user.id,
+        email: tokenJson.data.user.email,
+        companyId: tokenJson.data.user.companyId,
+        domainId: tokenJson.data.user.domainId,
+        root: tokenJson.data.user.root,
+        type: com.data.type ? com.data.type : "customer",
+      };
+      saveUser(userSession);
+
       setTitle("Logged In.");
+
+      // if (IsRemember) {
+      const token_data = {
+        companyId: com.data.id,
+        domainId: dom.data.id,
+        email: tokenJson.data.user.email,
+        accessToken: tokenJson.data.access_token,
+        userId: tokenJson.data.user.id,
+      };
+
+      await httpClient.post(uniDirServer.callback, token_data, {
+        withCredentials: true,
+        headers: headers,
+      });
+      // }
     }
   };
-  const onClicLogin = function (e) {
-    history.push("/login");
+  const onClicLogin = function () {
+    navigate("/login");
   };
   const loginButtonClass =
     " bg-gray-300 disabled:hover:cursor-not-allowed disabled:bg-gray-100 disabled:text-gray-400 enabled:transition enabled:transform enabled:hover:translate-x-1 enabled:hover:bg-gray-400 text-gray-800 font-bold py-2 px-4 rounded inline-flex items-center";
