@@ -3,13 +3,32 @@ import "../css/stepper.css";
 //import { TiTick } from "react-icons/ti";
 const Stepper = ({ steps, handleSubmit }) => {
   const [currentStep, setCurrentStep] = useState(1);
-  const [complete, setComplete] = useState(false);
-
+  const [complete, _] = useState(false);
+  const [error, setError] = useState(null);
   if (!steps) return <></>;
 
   const fixedButtonClass =
     "bg-gray-300 disabled:hover:cursor-not-allowed disabled:bg-gray-100 disabled:text-gray-400 enabled:transition enabled:transform enabled:hover:translate-x-1 enabled:hover:bg-gray-400 text-gray-800 font-bold py-2 px-4 rounded inline-flex items-center";
 
+  const verifyCurrentPage = (e) => {
+    if (steps[currentStep - 1].verify) {
+      const result = steps[currentStep - 1].verify();
+      if (result == "success") {
+        setError(null);
+      } else {
+        setError(result);
+        e.preventDefault();
+        return;
+      }
+    }
+
+    if (currentStep === steps.length) {
+      handleSubmit(e);
+    } else {
+      setCurrentStep((prev) => prev + 1);
+      e.preventDefault();
+    }
+  };
   return (
     <>
       <div className="flex justify-between">
@@ -40,6 +59,9 @@ const Stepper = ({ steps, handleSubmit }) => {
           </div>
         ))}
       </div>
+      {error ? (
+        <div className="w-full flex items-center text-red-500">{error}</div>
+      ) : null}
       <div className="w-full flex justify-between">
         {!complete && currentStep > 1 && (
           <button
@@ -55,12 +77,7 @@ const Stepper = ({ steps, handleSubmit }) => {
         {!complete && (
           <button
             className={fixedButtonClass}
-            onClick={(e) => {
-              currentStep === steps.length
-                ? handleSubmit(e)
-                : setCurrentStep((prev) => prev + 1);
-              e.preventDefault();
-            }}
+            onClick={(e) => verifyCurrentPage(e)}
           >
             {currentStep === steps.length ? "Finish" : "Next"}
           </button>
