@@ -22,8 +22,8 @@ function Toolbox({
       strokeWidth="2"
       stroke="currentColor"
       fill="none"
-      stroke-linecap="round"
-      stroke-linejoin="round"
+      strokeLinecap="round"
+      strokeLinejoin="round"
     >
       {" "}
       <path stroke="none" d="M0 0h24v24H0z" />{" "}
@@ -52,8 +52,8 @@ function Toolbox({
       strokeWidth="2"
       stroke="currentColor"
       fill="none"
-      stroke-linecap="round"
-      stroke-linejoin="round"
+      strokeLinecap="round"
+      strokeLinejoin="round"
     >
       {" "}
       <path stroke="none" d="M0 0h24v24H0z" />{" "}
