@@ -1,8 +1,8 @@
 const loginFields = [
   {
     labelText: "Email address",
-    labelFor: "email-address",
-    id: "email-address",
+    labelFor: "email",
+    id: "email",
     name: "email",
     type: "email",
     autoComplete: "email",
@@ -33,34 +33,39 @@ const signupFields = [
     placeholder: "Username",
   },
   {
-    labelText: "Email address",
-    labelFor: "email-address",
-    id: "email-address",
+    labelText: "Email",
+    labelFor: "email",
+    id: "email",
     name: "email",
-    type: "email",
+    type: "text",
     autoComplete: "email",
     isRequired: true,
-    placeholder: "Email address",
+    placeholder: "Email",
   },
   {
     labelText: "Password",
     labelFor: "password",
-    id: "password",
-    name: "password",
+    id: "authVerification",
+    name: "authVerification",
     type: "password",
-    autoComplete: "current-password",
+    autoComplete: "authVerification",
     isRequired: true,
     placeholder: "Password",
+    hiddenUpdate: true,
+    hiddenDisplay: true,
   },
   {
     labelText: "Confirm Password",
-    labelFor: "confirm-password",
-    id: "confirm-password",
-    name: "confirm-password",
+    labelFor: "confirmPassword",
+    id: "confirmPassword",
+    name: "confirmPassword",
     type: "password",
-    autoComplete: "confirm-password",
+    autoComplete: "confirmPassword",
     isRequired: true,
     placeholder: "Confirm Password",
+    database: false,
+    hiddenUpdate: true,
+    hiddenDisplay: true,
   },
 ];
 
