@@ -1,17 +1,23 @@
 import React from "react";
 import { useEffect, useState } from "react";
 import { apiPermissionScopesFields } from "../../constants/apiFields";
+import Toolbox from "../../component/Toolbox";
 import Pagination from "../../component/Pagination";
 import ApiPermissions from "./ApiPermissions";
 import apiApi from "../../api/api-api";
 import ButtonToolbox from "../../component/ButtonToolbox";
+import Modal from "../../component/Modal";
+import ItemField from "../../component/ItemField";
+import { generateString } from "../../utils/Utils";
 
 const fields = apiPermissionScopesFields;
+let fieldsState = {};
 
 function ApiPermissionScopePage({ mode, api }) {
   const pageDisplayCount = 10;
   const postDisplayCount = 15;
 
+  const [scopeMode, setScopeMode] = useState(mode);
   const [scopes, setScopes] = useState([]);
   const [scope, setScope] = useState({ permission: "", description: "" });
 
@@ -19,8 +25,18 @@ function ApiPermissionScopePage({ mode, api }) {
   const [currentPage, setCurrentPage] = useState(1);
   const [pageStart, setPageStart] = useState(1);
   const [pageEnd, setPageEnd] = useState(pageDisplayCount);
-
   const [postsPerPage] = useState(postDisplayCount);
+  const [checkedItems, setCheckedItems] = useState([]);
+  const [modalOpen, setModalOpen] = useState(false);
+  const [itemState, setItemState] = useState(
+    scopeMode === "new"
+      ? {
+          ...fieldsState,
+          permission: "api:read",
+        }
+      : fieldsState
+  );
+
   const indexOfLastPost = currentPage * postsPerPage;
   const indexOfFirstPost = indexOfLastPost - postsPerPage;
   const currentPosts =
@@ -31,15 +47,52 @@ function ApiPermissionScopePage({ mode, api }) {
     setPageEnd(endPage);
   };
 
-  const onClickNew = async function (e) {
-    await apiApi.createPermission(api.id, { ...scope, id: scope.permission });
+  const onClickAddScope = async function () {
+    setScopeMode("new");
     setScope({ permission: "", description: "" });
+    setItemState({ permission: "", description: "" });
+    setModalOpen(true);
+  };
+
+  const onCloseScopeModal = async function () {
+    setModalOpen(false);
+  };
+
+  const onClickNew = async function () {
+    const surfix = generateString(5);
+    await apiApi.createPermission(api.id, {
+      ...itemState,
+      id: `${itemState.permission}${surfix}`,
+    });
+    setScope({ permission: "", description: "" });
+    setItemState({ permission: "", description: "" });
     await getApiScopes();
+    setModalOpen(false);
+  };
+
+  const onClickEdit = async function (scope) {
+    setScopeMode("edit");
+    setScope(scope);
+    setItemState(scope);
+    setModalOpen(true);
+  };
+
+  const onClickApply = async function () {
+    await apiApi.updatePermission(api.id, itemState);
+    setScope({ permission: "", description: "" });
+    setItemState({ permission: "", description: "" });
+    await getApiScopes();
+    setModalOpen(false);
   };
 
   const onClickDel = async function (e) {
     //const api = await apiApi.get(company.id, domain.id, checkedItems);
     await apiApi.removePermission(api.id, e.id);
+    await getApiScopes();
+  };
+
+  const onClickDelMulti = async function (e) {
+    await apiApi.removePermission(api.id, checkedItems);
     await getApiScopes();
   };
 
@@ -53,6 +106,23 @@ function ApiPermissionScopePage({ mode, api }) {
     );
   };
 
+  const handleChange = (e) => {
+    const currentItem = fields.filter((f) => f.id === e.target.id)[0];
+    const itemValue =
+      e.target.value === "true" || e.target.value === "false"
+        ? e.target.checked
+        : currentItem.valueType !== undefined &&
+          currentItem.valueType === "array"
+        ? e.target.value.split(/\r\n|\n|\r/)
+        : e.target.value;
+
+    setItemState({ ...itemState, [e.target.id]: itemValue });
+  };
+
+  const handleCallback = (childCheckedItems) => {
+    setCheckedItems(childCheckedItems);
+  };
+
   useEffect(() => {
     setIsLoading(true);
     getApiScopes();
@@ -61,15 +131,15 @@ function ApiPermissionScopePage({ mode, api }) {
 
   // const delSvg = (
   //   <svg
-  //     class="w-4 h-4 mr-2"
+  //     className="w-4 h-4 mr-2"
   //     width="24"
   //     height="24"
   //     viewBox="0 0 24 24"
-  //     stroke-width="2"
+  //     strokeWidth="2"
   //     stroke="currentColor"
   //     fill="none"
-  //     stroke-linecap="round"
-  //     stroke-linejoin="round"
+  //     strokeLinecap="round"
+  //     strokeLinejoin="round"
   //   >
   //     {" "}
   //     <path stroke="none" d="M0 0h24v24H0z" />{" "}
@@ -85,15 +155,15 @@ function ApiPermissionScopePage({ mode, api }) {
 
   const primarySvg = (
     <svg
-      class="fill-current h-4 w-4 mr-2"
+      className="fill-current h-4 w-4 mr-2"
       fill="none"
       viewBox="0 0 24 24"
       stroke="currentColor"
     >
       <path
-        stroke-linecap="round"
-        stroke-linejoin="round"
-        stroke-width="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="2"
         d="M15 7a2 2 0 012 2m4 0a6 6 0 01-7.743 5.743L11 17H9v2H7v2H4a1 1 0 01-1-1v-2.586a1 1 0 01.293-.707l5.964-5.964A6 6 0 1121 9z"
       />
     </svg>
@@ -104,59 +174,18 @@ function ApiPermissionScopePage({ mode, api }) {
   return (
     <div className="col-span-full xl:col-span-6 bg-white dark:bg-slate-800 shadow-lg rounded-sm border border-slate-200 dark:border-slate-700">
       <header className="w-full px-5 py-4 border-b border-slate-100 dark:border-slate-700 relative inline-flex items-center justify-between">
-        {/* <h2 className="font-semibold text-slate-800 dark:text-slate-100">Manage Domain</h2> */}
-        <div className="space-y-4">
-          <label
-            for="permission"
-            className="text-pretty ms-2 text-sm font-medium text-gray-900 dark:text-gray-100 min-w-48 max-w-48 "
-          >
-            Permission
-          </label>
-          <br />
-          <input
-            id="permission"
-            value={scope.permission}
-            className="text-gray-800"
-            name="permission"
-            required={true}
-            placeholder="Permission"
-            onChange={(evt) =>
-              setScope({ ...scope, permission: evt.target.value })
-            }
-          />
-        </div>
         <div className="space-y-4 ml-10">
           <label
-            for="description"
-            className="text-pretty ms-2 text-sm font-medium text-gray-900 dark:text-gray-100 min-w-48 max-w-48 "
-          >
-            Description
-          </label>
-          <br />
-          <input
-            id="description"
-            value={scope.description}
-            className="text-gray-800"
-            name="description"
-            required={false}
-            placeholder="description"
-            onChange={(evt) =>
-              setScope({ ...scope, description: evt.target.value })
-            }
-          />
-        </div>
-        <div className="space-y-4 ml-10">
-          <label
-            for="description"
+            htmlFor="description"
             className="text-pretty ms-2 text-sm font-medium text-gray-900 dark:text-gray-100 min-w-48 max-w-48 "
           ></label>
           <br />
-          <ButtonToolbox
-            text="Add"
-            svg={primarySvg}
-            disabled={mode !== "edit"}
-            customClass={primaryButtonClass}
-            clickHandle={onClickNew}
+          <Toolbox
+            NewButtonLabel="Add"
+            onClickNew={onClickAddScope}
+            onClickDel={onClickDelMulti}
+            parentCallback={handleCallback}
+            disabledDel={checkedItems.length < 1}
           />
         </div>
       </header>
@@ -164,8 +193,10 @@ function ApiPermissionScopePage({ mode, api }) {
         {/* scopes */}
         <ApiPermissions
           scopes={currentPosts}
+          parentCallback={handleCallback}
           loading={isLoading}
           onClickDel={onClickDel}
+          onClickEdit={onClickEdit}
         />
         <Pagination
           postsPerPage={postsPerPage}
@@ -180,6 +211,38 @@ function ApiPermissionScopePage({ mode, api }) {
           pageStart={pageStart}
           pageEnd={pageEnd}
         />
+      </div>
+
+      <div className="flex items-center justify-center">
+        <Modal
+          hasCloseBtn={true}
+          isOpen={modalOpen}
+          onClose={onCloseScopeModal}
+          optionBtnLabel={scopeMode === "new" ? "Add" : "Apply"}
+          onOptionBtnClick={scopeMode === "new" ? onClickNew : onClickApply}
+        >
+          <div className="space-y-4">
+            {fields.map((field) =>
+              (field.hiddenUpdate || field.hiddenUpdate !== undefined) &&
+              scopeMode === "edit" ? (
+                <></>
+              ) : (
+                <ItemField
+                  key={field.name}
+                  item={scope}
+                  handleChange={handleChange}
+                  value={
+                    field.valueType === "array" && itemState[field.id]
+                      ? itemState[field.id].join("\r\n")
+                      : itemState[field.id]
+                  }
+                  field={field}
+                  mode={scopeMode}
+                />
+              )
+            )}
+          </div>
+        </Modal>
       </div>
     </div>
   );
