@@ -2,9 +2,10 @@ import React from "react";
 import { useState } from "react";
 import Checkbox from "../../component/Checkbox";
 
-function Apis({ apis, parentCallback, onClickView, onClickEdit }) {
+function Apis({ apis, parentCallback, onClickView, onClickEdit, customClass }) {
   const [checkedItems, setCheckedItems] = useState([]);
-
+  const [selected, setSelected] = useState(null);
+  console.log("apis items", apis);
   const handleChangeCheck = function (e) {
     const item = e.target.name;
     const isChecked = e.target.checked;
@@ -27,7 +28,7 @@ function Apis({ apis, parentCallback, onClickView, onClickEdit }) {
         <header className="text-xs uppercase text-slate-400 dark:text-slate-500 bg-slate-50 dark:bg-slate-700 dark:bg-opacity-50 rounded-sm font-semibold p-2">
           Apis
         </header>
-        <ul className="my-1">
+        <ul className={customClass ? customClass : "my-1 overflow-y-auto"}>
           {/* Item */}
           <li className="flex px-2">
             <div className="self-center mr-7"></div>
@@ -52,18 +53,27 @@ function Apis({ apis, parentCallback, onClickView, onClickEdit }) {
             apis.map((api) => {
               return (
                 <li
-                  className="flex px-2 hover:bg-indigo-200 dark:hover:bg-indigo-400 hover:cursor-pointer"
+                  className={
+                    selected === api.id
+                      ? "bg-indigo-400 flex px-2 hover:bg-indigo-200 dark:hover:bg-indigo-400 hover:cursor-pointer"
+                      : "flex px-2 hover:bg-indigo-200 dark:hover:bg-indigo-400 hover:cursor-pointer"
+                  }
                   key={api.id}
+                  onClick={() => setSelected(api.id)}
                 >
                   <div className="self-center mr-3">
-                    <Checkbox
-                      name={api.id}
-                      id={api.id}
-                      checked={
-                        checkedItems.filter((i) => i.id == api.id).length > 0
-                      }
-                      onChange={handleChangeCheck}
-                    />
+                    {onClickEdit ? (
+                      <Checkbox
+                        name={api.id}
+                        id={api.id}
+                        checked={
+                          checkedItems.filter((i) => i.id == api.id).length > 0
+                        }
+                        onChange={handleChangeCheck}
+                      />
+                    ) : (
+                      <></>
+                    )}
                     {/* <svg className="w-9 h-9 fill-current text-sky-50" viewBox="0 0 36 36">
                     <path d="M23 11v2.085c-2.841.401-4.41 2.462-5.8 4.315-1.449 1.932-2.7 3.6-5.2 3.6h-1v2h1c3.5 0 5.253-2.338 6.8-4.4 1.449-1.932 2.7-3.6 5.2-3.6h3l-4-4zM15.406 16.455c.066-.087.125-.162.194-.254.314-.419.656-.872 1.033-1.33C15.475 13.802 14.038 13 12 13h-1v2h1c1.471 0 2.505.586 3.406 1.455zM24 21c-1.471 0-2.505-.586-3.406-1.455-.066.087-.125.162-.194.254-.316.422-.656.873-1.028 1.328.959.878 2.108 1.573 3.628 1.788V25l4-4h-3z" />
                   </svg> */}
@@ -71,7 +81,7 @@ function Apis({ apis, parentCallback, onClickView, onClickEdit }) {
                   <div className="grow flex items-center border-b border-slate-100 dark:border-slate-700 text-sm py-2">
                     <div
                       className="grow flex"
-                      onClick={onClickView.bind(this, api)}
+                      onClick={onClickView ? onClickView.bind(this, api) : null}
                     >
                       <div className="self-center  w-1/6 min-w-48">
                         {api ? api.name : ""}
@@ -80,14 +90,18 @@ function Apis({ apis, parentCallback, onClickView, onClickEdit }) {
                         {api ? api.identifier : ""}
                       </div>
                     </div>
-                    <div
-                      className="shrink-0 self-end ml-2 w-1/6"
-                      onClick={onClickEdit.bind(this, api)}
-                    >
-                      <a className="font-medium text-indigo-500 hover:text-yellow-600 dark:hover:text-yellow-400">
-                        Edit
-                      </a>
-                    </div>
+                    {onClickEdit ? (
+                      <div
+                        className="shrink-0 self-end ml-2 w-1/6"
+                        onClick={onClickEdit.bind(this, api)}
+                      >
+                        <a className="font-medium text-indigo-500 hover:text-yellow-600 dark:hover:text-yellow-400">
+                          Edit
+                        </a>
+                      </div>
+                    ) : (
+                      <></>
+                    )}
                   </div>
                 </li>
               );
